backend/requirements.txt
appconfig==0.1
appdirs==1.4.4
Django==2.2.19
django-filter==21.1
djangorestframework==3.13.1
drf-extra-fields==3.1.1
gunicorn==20.1.0
Pillow==9.1.0
psycopg2==2.8.6
PyJWT==2.3.0
pytz==2022.1
sqlparse==0.4.2
